Initial dual mode

+++ extern.c
@@ -118,7 +118,7 @@ int quiet = 0;				/* Number of quiet turns */
 int vf_hit = 0;				/* Number of time flytrap has hit */
 
 int dnum;				/* Dungeon number */
-int seed;				/* Random number seed */
+uint64_t seed;				/* Random number seed */
 int e_levels[] = {
         10L,
 	20L,
